backend/requirements/test-requirements.txt
django-webtest==1.9.10
factory_boy==3.2.1
pytest==7.3.1
pytest-django==4.5.2
